.bmad-user-memory/modules/interactive-pdf-mapper/requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
httpx>=0.25.0
numpy>=1.24.0
